--- a/packages/grafana-ui/src/components/Select/utils.ts
+++ b/packages/grafana-ui/src/components/Select/utils.ts
@@ -20,19 +20,29 @@ export const cleanValue = (value: any, options: SelectOptions): SelectableValue[
   return [];
 };
 
+const _findSelectedValue = (value: string | number, option: SelectableValue): SelectableValue | null => {
+  if ('options' in option) {
+    let found = findSelectedValue(value, option.options);
+    if (found) {
+      return found;
+    }
+  } else if ('value' in option && option.value === value) {
+    return option;
+  }
+
+  return null;
+};
+
 /**
  * Find the label for a string|number value inside array of options or optgroups
  */
 export const findSelectedValue = (value: string | number, options: SelectOptions): SelectableValue | null => {
-  for (const option of options) {
-    if ('options' in option) {
-      let found = findSelectedValue(value, option.options);
-      if (found) {
-        return found;
-      }
-    } else if ('value' in option && option.value === value) {
-      return option;
+  if (Array.isArray(options)) {
+    for (const option of options) {
+      _findSelectedValue(value, option);
     }
+  } else {
+    _findSelectedValue(value, options);
   }
 
   return null;