--- a/packages/grafana-ui/tsconfig.json
+++ b/packages/grafana-ui/tsconfig.json
@@ -1,7 +1,6 @@
 {
   "compilerOptions": {
     "baseUrl": "node_modules/@types",
-    "declarationDir": "dist",
     "outDir": "compiled",
     "paths": {
       "@grafana/slate-react": ["slate-react"],