<!--

Thank you for sending a pull request! Here are some tips:

1. If this is your first time, please read our contribution guide at https://github.com/grafana/grafana/blob/master/CONTRIBUTING.md

2. Ensure you include and run the appropriate tests as part of your Pull Request.

3. In a new feature or configuration option, an update to the documentation is necessary. Everything related to the documentation is under the docs folder in the root of the repository.

4. If the Pull Request is a work in progress, make use of GitHub's "Draft PR" feature and mark it as such.

5. If you can not merge your Pull Request due to a merge conflict, Rebase it. This gets it in sync with the master branch.

6. Name your PR as "<FeatureArea>: Describe your change", e.g. Alerting: Prevent race condition. If it's a fix or feature relevant for the changelog describe the user impact in the title. The PR title is used to auto-generate the changelog for issues marked with the "add to changelog" label.

-->

**What this PR does / why we need it**:

I wanted to improve Graph component in our fork regarding to https://github.com/grafana/grafana/issues/26523 and found out that developing using current storybook setup veeeery slow.. So I decided to improve build a little bit:

- upgrade loaders versions (`expose-loader` improve perf a lot)
- use react-docgen-typescript-loader only for production

**Which issue(s) this PR fixes**:

no issue
<!--

* Automatically closes linked issue when the Pull Request is merged.

Usage: "Fixes #<issue number>", or "Fixes (paste link of issue)"

-->

Fixes #

**Special notes for your reviewer**:

|       | cold start | incremental build |
|----|-----------|------------------|
| `master` https://github.com/grafana/grafana/commit/398bc045fc0094d9a1482f040d7ef334c6a8034b | 1.73 min for manager and 3.1 min for preview | 5-12s depending on changes |
|  PR           | 56s , 1.38min | 5-11.5s depending on changes |
| PR + transpileOnly | 25s, 45s | <4.5s |

This PR improves mostly only cold start. To improve incremental build I strongly recommend to move out of `ts-loader` and use babel instead or solve `ts-loader` issues. Also shipping es6+ in dev allow to use `options.devtool = false`

Profile of HMR update below
<img width="1416" alt="Снимок экрана 2020-08-08 в 1 00 15" src="https://user-images.githubusercontent.com/10380560/89692299-f1a5c380-d913-11ea-881f-3531d5785639.png">