@aknuds1 2020-08-08
The front-end tests [fail](https://drone.grafana.net/grafana/grafana/1070/1/12).