@vankop 2020-08-08
> The front-end tests fail.

hm.. whatever, I have reverted utils changes