@torkelo 2020-08-09
Awesome work!  Big thanks from the frontend platform squad! :heart: we will look at this asap in the coming 2-3 days