@vankop 2020-08-14
I am busy right now, feel free to pick it up + will be great to see incremental build improvments