@aknuds1 2020-08-14
Unfortunately I'm not involved in the front-end development, so I don't know how to tweak the build :/