@zoltanbedi 2020-08-14
Thanks for contributing @vankop. Let me share my 2c here. I think we should upgrade storybook to v6 first and then revisit the build performance. I know that the main problem is the props table generator because you disabled it for development it means that the ci is still going to be slow. All in all I like the idea to make this faster but I think we should upgrade to v6 first.