@mckn 2020-08-28
A quick update: @peterholmberg is currently looking into upgrading storybook 6 and we will see what bits and pieces from this PR we can apply after that is done.

I love this initiative and great that we could speed up things some more.