@mckn 2020-09-23
As I mentioned earlier we are looking into upgrading Storybook to version 6 so we will close this PR. We would love to get some help of speeding up the upgraded version of Storybook once that is in place.

Issue: https://github.com/grafana/grafana/issues/25312

Thanks again for opening this PR 🙌